[refactor]: 공연 서비스 로직 분리
